requirements.txt
aiogram==3.24.0
ccxt==4.5.32
pandas>=2.1
numpy>=1.26
pyarrow>=14.0
matplotlib>=3.8
mplfinance>=0.12.10b0
httpx>=0.25
pydantic>=2.5
python-dotenv>=1.0
loguru>=0.7
apscheduler>=3.10
